Implement better chaining and config
### Chaining
This commit adds a mechanism that should make chaining work better "out-of-the-box". When chaining several operations (especially when using custom labels) you would need to tell the operations later in the chain about these custom labels in `ignore_keys` so that they can be ignored. A mechanism has been added now that stores the `ignored_keys` in the `df.attrs`. A decorator checks if there are any `ignored_keys` stored and if so sets them and finally also adds any labels that need to be ignored when performing another flatbread operation. See "chaining.py" for the implementation.

### Config
This commit also adds a way to set defaults from a "config.json" file. This mechanism has also been implemented with a decorator. It will check if any keywords that have defaults have not been set or are `None` and if so set them. A drawback of this is that downstream the defaults cannot be set but have to be `None` otherwise they will not be overridden by the defaults at the source.

### Other changes
* This commit also reorganizes the code. The `DataFrame` and `Series` accessors have been put in their own  modules.
* An option has been added to return the `Series`/`DataFrame` as percentages (instead of adding percentages to the data).
* Subaggregation has been added for the `Series`. Because the implementation for `Series` and `DataFrame` is mostly the same, both call the same underlying `_subagg_implementation()`.