Fix issues with the display
When section_levels were not set in the config then this would create incorrect columnspans in the header for the column group labels. This has been fixed in wc-simple-table as well but here the template should no longer add the attribute if it was not in the config.

This same philosophy has been applied to all the attributes. Moreover, the template has been changed to remove most of the unnecessary whitespace.

The json serializer has been moved from the template handler to the table spec where it belongs. Moreover, an issue has been fixed with the serializer not converting NaN to null. This is now handled at the level of converting the values to a list.