Add option to specify a list of keys when adding level
Previously it was only possible to add one singular value to all index keys. With this it is also possible to specify a list of values to be added to a level of the multiindex. Currently, the number of values supplied must be equal to the length of the index.